Move CLI time utilities into shared reporting boundary
Relocated time helpers from src/cli/shared/time-utils.js to src/shared/reporting/time-utils.js so the duration formatting and verbose logging utilities live alongside createMetricsTracker and can be reused across CLI and plugin reporting features.
