+++ src/cli/features/manual/utils.js
@@ -10,7 +10,7 @@ import {
     parseJsonWithContext,
     toTrimmedString
 } from "../../shared/dependencies.js";
-import { formatDuration } from "../../shared/time-utils.js";
+import { formatDuration } from "../../shared/reporting.js";
 import { formatBytes } from "../../runtime-options/byte-format.js";
 import { writeManualFile } from "./file-helpers.js";
 import {
