+++ src/shared/reporting/index.js
@@ -1,1 +1,6 @@
 export { createMetricsTracker } from "./metrics.js";
+export {
+    formatDuration,
+    createVerboseDurationLogger,
+    timeSync
+} from "./time-utils.js";
