+++ src/shared/reporting/time-utils.js
